Switch to use stacking flag
