+++ src/accel/LocalOpticalGenOffload.cc
@@ -15,10 +15,7 @@
 #include "corecel/sys/ActionRegistryOutput.hh"
 #include "corecel/sys/Device.hh"
 #include "corecel/sys/ScopedProfiling.hh"
-#include "geocel/GeantGeoParams.hh"
 #include "geocel/GeantUtils.hh"
-#include "geocel/GeoOpticalIdMap.hh"
-#include "geocel/g4/Convert.hh"
 #include "celeritas/global/CoreParams.hh"
 #include "celeritas/optical/CoreParams.hh"
 #include "celeritas/optical/CoreState.hh"
@@ -143,52 +140,6 @@ void LocalOpticalGenOffload::Push(optical::GeneratorDistributionData const& data
     {
         this->Flush();
     }
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Construct a generator distribution from a \c G4Step and buffer it.
- */
-void LocalOpticalGenOffload::Push(G4Step const& step,
-                                  GeneratorType gen_type,
-                                  size_type num_photons)
-{
-    CELER_EXPECT(num_photons > 0);
-
-    auto geant_geo = celeritas::global_geant_geo().lock();
-    CELER_VALIDATE(geant_geo, << "global Geant4 geometry is not loaded");
-
-    auto* pre_step = step.GetPreStepPoint();
-    auto* post_step = step.GetPostStepPoint();
-    CELER_ASSERT(pre_step && post_step);
-
-    optical::GeneratorDistributionData data;
-    data.step_length
-        = native_from_geant<lengthunits::ClhepLength>(step.GetStepLength());
-    data.charge = units::ElementaryCharge{
-        static_cast<real_type>(post_step->GetCharge())};
-    data.type = gen_type;
-    data.num_photons = num_photons;
-
-    auto load_step_point = [](G4StepPoint const* p) {
-        optical::GeneratorStepData data;
-        data.speed = units::LightSpeed(p->GetBeta());
-        data.time = native_from_geant<units::ClhepTime>(p->GetGlobalTime());
-        data.pos = native_from_geant<lengthunits::ClhepLength, real_type>(
-            p->GetPosition());
-        return data;
-    };
-
-    data.points[StepPoint::pre] = load_step_point(pre_step);
-    data.points[StepPoint::post] = load_step_point(post_step);
-
-    auto* g4mat = pre_step->GetMaterial();
-    CELER_ASSERT(g4mat);
-    data.material
-        = (*geant_geo->geo_optical_id_map())[geant_geo->geant_to_id(*g4mat)];
-
-    CELER_ASSERT(data);
-    this->Push(data);
 }
 
 //---------------------------------------------------------------------------//
