+++ src/accel/LocalOpticalGenOffload.hh
@@ -16,7 +16,6 @@
 #include "LocalOffloadInterface.hh"
 
 class G4EventManager;
-class G4Step;
 
 namespace celeritas
 {
@@ -76,9 +75,6 @@ class LocalOpticalGenOffload final : public LocalOffloadInterface
 
     // Offload optical distribution data to Celeritas
     void Push(DistributionData const&);
-
-    // Construct and offload optical distribution data to Celeritas from a step
-    void Push(G4Step const&, GeneratorType, size_type num_photons);
 
     //! Whether the class instance is initialized
     explicit operator bool() const { return this->Initialized(); }
