+++ src/accel/gen/G4OffloadUtils.hh
@@ -1,0 +1,26 @@
+//------------------------------- -*- C++ -*- -------------------------------//
+// Copyright Celeritas contributors: see top-level COPYRIGHT file for details
+// SPDX-License-Identifier: (Apache-2.0 OR MIT)
+//---------------------------------------------------------------------------//
+//! \file accel/gen/G4OffloadUtils.hh
+//---------------------------------------------------------------------------//
+#pragma once
+
+class G4Step;
+
+namespace celeritas
+{
+namespace optical
+{
+struct GeneratorDistributionData;
+}
+
+//---------------------------------------------------------------------------//
+// FREE FUNCTIONS
+//---------------------------------------------------------------------------//
+
+// Populate a \c GeneratorDistributionData with \c G4Step data
+optical::GeneratorDistributionData distribution_from_step(G4Step const&);
+
+//---------------------------------------------------------------------------//
+}  // namespace celeritas
